+++ docs/CHANGELOG.md
@@ -14,9 +14,12 @@ The entire project (i.e., the GraphQL API, REST API, and server) uses the same v
 - `Query.searchGroupChatUsers`
 - `type UpdatedPollMessage`
 - `union ReadMessageResult`
+- `type InvalidBroadcast`
 
 ### Changed
 
+- Add `type InvalidBroadcast` to `union CreateTextMessageResult`.
+- Return a `type InvalidBroadcast` instead of an authorization error in `Mutation.createTextMessage`.
 - Return a `union ReadMessageResult` instead of a `interface Message` from `Query.readMessage`.
 - Rename the `title` field to `question` in `type Poll` and `input PollInput`.
 - Rename `Mutation.setInvitability` to `Mutation.setPublicity`.